Refactoring the Traject indexing configuration to use recursive component indexing within colletions; Scaling down the EAD fixtures in order to improve the speed of test suites
