+++ .rubocop.yml
@@ -12,6 +12,10 @@ AllCops:
     - 'node_modules/**/*'
     - 'spec/support/**/*'
     - 'tmp/**/*'
+
+Metrics/AbcSize:
+  Exclude:
+    - 'lib/pulfalight/traject/ead2_component_config.rb'
 Metrics/BlockLength:
   Exclude:
     - 'spec/**/*'
@@ -27,9 +31,20 @@ Metrics/MethodLength:
     - 'lib/tasks/pulfalight.rake'
     - 'app/models/solr_document.rb'
     - 'app/controllers/catalog_controller.rb'
+    - 'lib/pulfalight/traject/ead2_component_config.rb'
 Naming/PredicateName:
   Exclude:
     - 'app/models/solr_document.rb'
+Metrics/CyclomaticComplexity:
+  Exclude:
+    - 'lib/pulfalight/traject/ead2_component_config.rb'
+Metrics/ModuleLength:
+  Exclude:
+    - 'lib/pulfalight/traject/ead2_component_config.rb'
+Metrics/PerceivedComplexity:
+  Exclude:
+    - 'lib/pulfalight/traject/ead2_component_config.rb'
+
 RSpec/ExampleLength:
   Exclude:
     - 'spec/features/traject/ead2_indexing_spec.rb'
