+++ lib/pulfalight/traject/ead2_indexing.rb
@@ -1,0 +1,83 @@
+# frozen_string_literal: true
+
+module Pulfalight
+  module Ead2Indexing
+    NAME_ELEMENTS = %w[corpname famname name persname].freeze
+
+    SEARCHABLE_NOTES_FIELDS = %w[
+      accessrestrict
+      accruals
+      altformavail
+      appraisal
+      arrangement
+      bibliography
+      bioghist
+      custodhist
+      fileplan
+      note
+      odd
+      originalsloc
+      otherfindaid
+      phystech
+      prefercite
+      processinfo
+      relatedmaterial
+      scopecontent
+      separatedmaterial
+      userestrict
+    ].freeze
+
+    DID_SEARCHABLE_NOTES_FIELDS = %w[
+      abstract
+      materialspec
+      physloc
+    ].freeze
+
+    def configure_before
+      settings do
+        provide "reader_class_name", "Arclight::Traject::NokogiriNamespacelessReader"
+        provide "solr_writer.commit_on_close", "true"
+        provide "repository", ENV["REPOSITORY_ID"]
+        provide "logger", Logger.new($stderr)
+      end
+
+      each_record do |_record, context|
+        next unless settings["repository"]
+
+        repository = Arclight::Repository.find_by(
+          slug: settings["repository"]
+        )
+
+        context.clipboard[:repository] = repository.name unless repository.nil?
+      end
+    end
+
+    def configure_after
+      each_record do |_record, context|
+        context.output_hash["components"] &&= context.output_hash["components"].select { |c| c.keys.any? }
+      end
+    end
+
+    def build_component_indexer(context)
+      config_file_path = Rails.root.join("lib", "pulfalight", "traject", "ead2_component_config.rb")
+      indexer_settings = { parent: context }
+      Traject::Indexer::NokogiriIndexer.new(indexer_settings).tap do |i|
+        i.load_config_file(config_file_path)
+      end
+    end
+
+    ##
+    # Used for evaluating xpath components to find
+    class NokogiriXpathExtensions
+      # rubocop:disable Naming/PredicateName, Style/FormatString
+      def is_component(node_set)
+        node_set.find_all do |node|
+          component_elements = (1..12).map { |i| "c#{'%02d' % i}" }
+          component_elements.push "c"
+          component_elements.include? node.name
+        end
+      end
+      # rubocop:enable Naming/PredicateName, Style/FormatString
+    end
+  end
+end
